# bruno-collection/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Faz o Simplesebol API",
    "version": "1.0.0",
    "description": "API REST para gerenciamento de jogos de futebol com sistema de pagamentos e controle de jogadores",
    "contact": {
      "name": "API Support",
      "email": "support@botfutebol.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento"
    },
    {
      "url": "https://api.botfutebol.com",
      "description": "Servidor de Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GameResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do jogo",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "description": "Nome/título do jogo",
            "example": "⚽ CAMPO DO VIANA"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Data do jogo (YYYY-MM-DD)",
            "example": "2024-12-25"
          },
          "time": {
            "type": "string",
            "description": "Horário do jogo (HH:MM)",
            "example": "20:30"
          },
          "maxPlayers": {
            "type": "integer",
            "description": "Número máximo de jogadores",
            "example": 16
          },
          "currentPlayers": {
            "type": "integer",
            "description": "Número atual de jogadores confirmados",
            "example": 12
          },
          "pricePerPlayer": {
            "type": "number",
            "description": "Preço por jogador em reais",
            "example": 14
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "open",
              "closed",
              "finished",
              "cancelled"
            ],
            "description": "Status atual do jogo",
            "example": "open"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do jogo",
            "example": "2024-12-20T10:00:00.000Z"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/games": {
      "get": {
        "summary": "Lista todos os jogos",
        "description": "Retorna uma lista paginada de jogos com filtros opcionais",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "open",
                "closed",
                "finished",
                "cancelled"
              ]
            },
            "description": "Filtrar por status do jogo"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por tipo de jogo"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por título ou localização"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número da página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Itens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de jogos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "games": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GameResponseDto"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total de jogos"
                    },
                    "page": {
                      "type": "integer",
                      "description": "Página atual"
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total de páginas"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um novo jogo",
        "description": "Cria um novo jogo agendado (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Jogo criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão de admin"
          }
        }
      }
    },
...
    "/api/games/{gameId}": {
      "get": {
        "summary": "Obtém detalhes de um jogo",
        "description": "Retorna informações detalhadas de um jogo específico incluindo jogadores, lista de espera e resumo financeiro",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do jogo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Jogo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualiza um jogo",
        "description": "Atualiza informações de um jogo existente (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jogo atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão de admin"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Cancela um jogo",
        "description": "Cancela um jogo agendado (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Jogo cancelado com sucesso"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/status": {
      "put": {
        "summary": "Atualiza status do jogo",
        "description": "Atualiza apenas o status de um jogo (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "scheduled",
                      "open",
                      "closed",
                      "finished",
                      "cancelled"
                    ],
                    "description": "Novo status do jogo",
                    "example": "open"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Status inválido"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão de admin"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/close": {
      "post": {
        "summary": "Fecha um jogo",
        "description": "Fecha um jogo e processa pagamentos/débitos (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Jogo fechado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão de admin"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/send-reminder": {
      "post": {
        "summary": "Envia lembrete do jogo",
        "description": "Envia um lembrete via WhatsApp para o grupo do jogo (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "responses": {
          "204": {
            "description": "Lembrete enviado com sucesso"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão de admin"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/export": {
      "get": {
        "summary": "Exporta lista de jogadores em CSV",
        "description": "Exporta a lista de jogadores do jogo em formato CSV (requer autenticação de admin)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo CSV gerado com sucesso",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão de admin"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/players": {
      "post": {
        "summary": "Adiciona um jogador ou convidado ao jogo",
        "description": "Adiciona um novo jogador à lista do jogo (requer autenticação).\n\n**Para adicionar jogador normal:**\n- Envie `phone` e `name`\n\n**Para adicionar convidado:**\n- Envie `phone` (de quem convida), `name` (de quem convida), e `guestName` (nome do convidado)\n- O convidado será adicionado com formato: \"NomeConvidado (conv. NomeQuemConvida)\"\n",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlayerToGameDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Jogador adicionado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autenticado"
          },
          "404": {
            "description": "Jogo não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/players/{playerId}": {
      "delete": {
        "summary": "Remove um jogador do jogo",
        "description": "Remove um jogador da lista do jogo (requer autenticação)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogo"
          },
          {
            "in": "path",
            "name": "playerId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do jogador"
          }
        ],
        "responses": {
          "204": {
            "description": "Jogador removido com sucesso"
          },
          "401": {
            "description": "Não autenticado"
          },
          "404": {
            "description": "Jogo ou jogador não encontrado"
          }
        }
      }
    },
    "/api/games/{gameId}/remove-guest": {
      "post": {
        "summary": "Remove um convidado por slot (Preciso)",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "slot"
                ],
                "properties": {
                  "slot": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso"
          }
        }
      }
    },
    "/api/games/{gameId}/remove-player": {
      "post": {
        "summary": "Remove um jogador por ID (Preciso)",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso"
          }
        }
      }
    },
...
  },
...
}